fixes schedule HTTP verb support, breaks into modules
