+++ lambda/schedule/postSchedule.js
@@ -1,28 +1,8 @@
 'use strict';
 
-let UpdateExpression = "SET"
-let ExpressionAttributeValues = {};
-
-module.exports.updateSchedule = (obj, table) => {
-  const { DeviceId, ScheduleId, ...input } = obj;
-
-  Object.keys(input).forEach((key, idx, arr) => {
-    let newKey = `:${key}`;
-    UpdateExpression += ` ${key} = ${newKey}`
-    if (idx + 1 !== arr.length) UpdateExpression += ",";
-    ExpressionAttributeValues[newKey] = obj[key];
-  });
-
-  const params = {
-    UpdateExpression,
-    ExpressionAttributeValues,
-    "Key": {
-      "DeviceId": DeviceId,
-      "ScheduleId": ScheduleId
-    },
-    "TableName": table
+module.exports.createSchedule = (obj, table) => {
+  return {
+    Item: obj,
+    TableName: table
   };
-
-  console.log(params);
-  return params;
 };
